Move the NIST assembly task from core to contrib
The task is a contributed task family, so it belongs under contrib
rather than core. Nothing in core imported it, so the dependency
direction stated in isaaclab_tasks (core never depends on contrib) is
preserved.

Name the package after the NIST taskboard rather than the task: contrib
already ships an unrelated factory package, and the old multi_task name
described the directory's history rather than its contents. The gym id
IsaacContrib-Factory-Franka is unchanged.
