+++ source/isaaclab_tasks/changelog.d/zhengyuz-factory-newton.rst
@@ -2,5 +2,5 @@ Added
 ^^^^^
 
 * Added the ``IsaacContrib-Factory-Franka`` contact-rich assembly task under
-  :mod:`isaaclab_tasks.core.multi_task`, covering NIST taskboard nut-threading and
+  :mod:`isaaclab_tasks.contrib.nist`, covering NIST taskboard nut-threading and
   peg insertion with a Franka Panda on the Newton backend.
