+++ source/isaaclab_tasks/isaaclab_tasks/contrib/nist/assembly_keypoints.py
