+++ source/isaaclab_tasks/isaaclab_tasks/contrib/nist/factory_assets_cfg.py
