+++ source/isaaclab_tasks/isaaclab_tasks/contrib/nist/factory_scenes_cfg.py
