+++ source/isaaclab_tasks/isaaclab_tasks/contrib/nist/utils/rigid_object_hasher.py
